+++ src/main/java/org/songeun/petdongne_server/review/infrastructure/ResidenceReviewRepository.java
@@ -1,6 +1,6 @@
 package org.songeun.petdongne_server.review.infrastructure;
 
-import org.songeun.petdongne_server.review.domain.ResidenceReview;
+import org.songeun.petdongne_server.review.domain.entity.ResidenceReview;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
