+++ src/main/java/org/songeun/petdongne_server/review/infrastructure/ReviewPhotoRepository.java
@@ -1,6 +1,6 @@
 package org.songeun.petdongne_server.review.infrastructure;
 
-import org.songeun.petdongne_server.review.domain.ReviewPhoto;
+import org.songeun.petdongne_server.review.domain.entity.ReviewPhoto;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
