+++ src/main/java/org/songeun/petdongne_server/survey/domain/SurveyResponse.java
@@ -4,7 +4,7 @@ import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
-import org.songeun.petdongne_server.review.domain.ResidenceReview;
+import org.songeun.petdongne_server.review.domain.entity.ResidenceReview;
 import org.songeun.petdongne_server.global.common.BaseEntity;
 
 @Entity
